Move methods.ts to api/methods.ts
