+++ src/index.ts
@@ -15,10 +15,10 @@ export * from './api/Errors';
 export * from './api/IntentMetadata';
 export * from './api/IntentResolution';
 export * from './api/Listener';
+export * from './api/methods';
 export * from './context/ContextType';
 export * from './context/ContextTypes';
 export * from './intents/Intents';
-export * from './methods';
 
 declare global {
   interface Window {
